Create test-send-receiver test client
